# src/main/resources/openapi_napolya.yaml
openapi: 3.0.3
info:
  title: API na_polya 3.0
  version: 1.0.11
  description: Все необходимые API для сайта На_поля
...
paths:
...
  /teams:
    get:
      tags:
        - Team
      summary: Получить список названий команд
      responses:
        '200':
          description: Успешный ответ с массивом названий команд
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: Список названий команд
                  example: ["Команда Победителей", "Лига Атлетов"]
    post:
      tags:
        - Team
      summary: Создать новую команду
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
      responses:
        '201':
          description: Команда успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
  /teams/{id}:
    get:
      tags:
        - Team
      summary: Получить информацию о команде
      parameters:
        - name: id
          in: path
          required: true
          description: ID команды
          schema:
            type: integer
      responses:
        '200':
          description: Информация о команде
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          description: Команда не найдена

    put:
      tags:
        - Team
      summary: Обновить информацию о команде
      parameters:
        - name: id
          in: path
          required: true
          description: ID команды
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
      responses:
        '200':
          description: Команда успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          description: Команда не найдена
    
    delete:
      tags:
        - Team
      summary: Удалить команду
      parameters:
        - name: id
          in: path
          required: true
          description: ID команды
          schema:
            type: integer
      responses:
        '204':
          description: Команда успешно удалена
        '404':
          description: Команда не найдена
...
components:
  schemas:
...
    Team: # Определение сущности "Команда"
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор команды
          example: 1
        name:
          type: string
          description: Название команды
          example: "Команда Победителей"
        players_ids:
          type: array
          description: id игроков, входящих в состав команды
          items:
            type: integer
          example: [2, 3, 4]
        game_id:
          type: integer
          description: id игры, в которой участвует команда
          example: 1
        captain_id:
          type: integer
          description: id капитана команды
          example: 2
      
      required:
        - id
        - name
        - game_id
        - captain_id
...